refactor(about): extract utils and phrases for coverage and duplication
- Extract parseCSVLine, getRandomThankYou, getFontSize, generateGradientColor,
  cleanName to aboutUtils.ts (exported, testable)
- Extract THANK_YOU_PHRASES to thankYouPhrases.ts (data-only module)
- Add aboutUtils.test.ts with 24 unit tests covering all utilities
- Add thankYouPhrases.ts to sonar.cpd.exclusions (translation data)
- About.tsx reduced from 764 to 257 lines
