+++ sonar-project.properties
@@ -20,7 +20,8 @@ sonar.exclusions=\
   apps/frontend/src/api/generated/**
 
 sonar.cpd.exclusions=\
-  apps/frontend/src/api/generated/**
+  apps/frontend/src/api/generated/**,\
+  apps/frontend/src/pages/thankYouPhrases.ts
 
 sonar.test.exclusions=\
   **/node_modules/**,\
